tear out import/export
