+++ cmd/notary/keys_test.go
@@ -533,93 +533,6 @@ func TestChangeKeyPassphraseNonexistentID(t *testing.T) {
 	require.Contains(t, err.Error(), "could not retrieve local key for key ID provided")
 }
 
-func TestKeyImportMismatchingRoles(t *testing.T) {
-	setUp(t)
-	k := &keyCommander{
-		configGetter:   func() (*viper.Viper, error) { return viper.New(), nil },
-		getRetriever:   func() notary.PassRetriever { return passphrase.ConstantRetriever("pass") },
-		keysImportRole: "targets",
-	}
-	tempFileName := generateTempTestKeyFile(t, "snapshot")
-	defer os.Remove(tempFileName)
-
-	err := k.keysImport(&cobra.Command{}, []string{tempFileName})
-	require.Error(t, err)
-	require.Contains(t, err.Error(), "does not match role")
-}
-
-func TestKeyImportNoGUNForTargetsPEM(t *testing.T) {
-	setUp(t)
-	k := &keyCommander{
-		configGetter: func() (*viper.Viper, error) { return viper.New(), nil },
-		getRetriever: func() notary.PassRetriever { return passphrase.ConstantRetriever("pass") },
-	}
-	tempFileName := generateTempTestKeyFile(t, "targets")
-	defer os.Remove(tempFileName)
-
-	err := k.keysImport(&cobra.Command{}, []string{tempFileName})
-	require.Error(t, err)
-	require.Contains(t, err.Error(), "Must specify GUN")
-}
-
-func TestKeyImportNoGUNForSnapshotPEM(t *testing.T) {
-	setUp(t)
-	k := &keyCommander{
-		configGetter: func() (*viper.Viper, error) { return viper.New(), nil },
-		getRetriever: func() notary.PassRetriever { return passphrase.ConstantRetriever("pass") },
-	}
-	tempFileName := generateTempTestKeyFile(t, "snapshot")
-	defer os.Remove(tempFileName)
-
-	err := k.keysImport(&cobra.Command{}, []string{tempFileName})
-	require.Error(t, err)
-	require.Contains(t, err.Error(), "Must specify GUN")
-}
-
-func TestKeyImportNoGUNForTargetsFlag(t *testing.T) {
-	setUp(t)
-	k := &keyCommander{
-		configGetter:   func() (*viper.Viper, error) { return viper.New(), nil },
-		getRetriever:   func() notary.PassRetriever { return passphrase.ConstantRetriever("pass") },
-		keysImportRole: "targets",
-	}
-	tempFileName := generateTempTestKeyFile(t, "")
-	defer os.Remove(tempFileName)
-
-	err := k.keysImport(&cobra.Command{}, []string{tempFileName})
-	require.Error(t, err)
-	require.Contains(t, err.Error(), "Must specify GUN")
-}
-
-func TestKeyImportNoGUNForSnapshotFlag(t *testing.T) {
-	setUp(t)
-	k := &keyCommander{
-		configGetter:   func() (*viper.Viper, error) { return viper.New(), nil },
-		getRetriever:   func() notary.PassRetriever { return passphrase.ConstantRetriever("pass") },
-		keysImportRole: "snapshot",
-	}
-	tempFileName := generateTempTestKeyFile(t, "")
-	defer os.Remove(tempFileName)
-
-	err := k.keysImport(&cobra.Command{}, []string{tempFileName})
-	require.Error(t, err)
-	require.Contains(t, err.Error(), "Must specify GUN")
-}
-
-func TestKeyImportNoRole(t *testing.T) {
-	setUp(t)
-	k := &keyCommander{
-		configGetter: func() (*viper.Viper, error) { return viper.New(), nil },
-		getRetriever: func() notary.PassRetriever { return passphrase.ConstantRetriever("pass") },
-	}
-	tempFileName := generateTempTestKeyFile(t, "")
-	defer os.Remove(tempFileName)
-
-	err := k.keysImport(&cobra.Command{}, []string{tempFileName})
-	require.Error(t, err)
-	require.Contains(t, err.Error(), "Could not infer role, and no role was specified for key")
-}
-
 func generateTempTestKeyFile(t *testing.T, role string) string {
 	setUp(t)
 	privKey, err := trustmanager.GenerateECDSAKey(rand.Reader)
