+++ cmd/notary/main_test.go
@@ -159,10 +159,6 @@ var exampleValidCommands = []string{
 	"key list",
 	"key rotate repo snapshot",
 	"key generate rsa",
-	"key backup tempfile.zip",
-	"key export e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855 backup.pem",
-	"key restore tempfile.zip",
-	"key import backup.pem",
 	"key remove e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855",
 	"key passwd e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855",
 	"delegation list repo",
