+++ trustmanager/keyfilestore_test.go
@@ -660,75 +660,3 @@ func TestKeysAreCached(t *testing.T) {
 	}
 	require.Equal(t, 2, numTimesCalled, "numTimesCalled should be 2 -- no additional call to passphraseRetriever")
 }
-
-// Exporting a key is successful (it is a valid key)
-func TestKeyFileStoreExportSuccess(t *testing.T) {
-	// Generate a new Private Key
-	privKey, err := GenerateECDSAKey(rand.Reader)
-	require.NoError(t, err)
-
-	// Temporary directory where test files will be created
-	tempBaseDir, err := ioutil.TempDir("", "notary-test-")
-	require.NoError(t, err)
-	defer os.RemoveAll(tempBaseDir)
-
-	// Create our FileStore and add the key
-	store, err := NewKeyFileStore(tempBaseDir, passphraseRetriever)
-	require.NoError(t, err)
-	err = store.AddKey(KeyInfo{Role: data.CanonicalRootRole, Gun: ""}, privKey)
-	require.NoError(t, err)
-
-	assertExportKeySuccess(t, store, privKey)
-}
-
-// Exporting a key that doesn't exist fails (it is a valid key)
-func TestKeyFileStoreExportNonExistantFailure(t *testing.T) {
-	// Temporary directory where test files will be created
-	tempBaseDir, err := ioutil.TempDir("", "notary-test-")
-	require.NoError(t, err)
-	defer os.RemoveAll(tempBaseDir)
-
-	// Create empty FileStore
-	store, err := NewKeyFileStore(tempBaseDir, passphraseRetriever)
-	require.NoError(t, err)
-
-	_, err = store.ExportKey("12345")
-	require.Error(t, err)
-}
-
-// Exporting a key is successful (it is a valid key)
-func TestKeyMemoryStoreExportSuccess(t *testing.T) {
-	// Generate a new Private Key
-	privKey, err := GenerateECDSAKey(rand.Reader)
-	require.NoError(t, err)
-
-	// Create our MemoryStore and add key to it
-	store := NewKeyMemoryStore(passphraseRetriever)
-	require.NoError(t, err)
-	err = store.AddKey(KeyInfo{Role: data.CanonicalRootRole, Gun: ""}, privKey)
-	require.NoError(t, err)
-
-	assertExportKeySuccess(t, store, privKey)
-}
-
-// Exporting a key that doesn't exist fails (it is a valid key)
-func TestKeyMemoryStoreExportNonExistantFailure(t *testing.T) {
-	store := NewKeyMemoryStore(passphraseRetriever)
-	_, err := store.ExportKey("12345")
-	require.Error(t, err)
-}
-
-// Given a keystore and expected key that is in the store, export the key
-// and assert that the exported key is the same and encrypted with the right
-// password.
-func assertExportKeySuccess(
-	t *testing.T, s KeyStore, expectedKey data.PrivateKey) {
-
-	pemBytes, err := s.ExportKey(expectedKey.ID())
-	require.NoError(t, err)
-
-	reparsedKey, err := ParsePEMPrivateKey(pemBytes, cannedPassphrase)
-	require.NoError(t, err)
-	require.Equal(t, expectedKey.Private(), reparsedKey.Private())
-	require.Equal(t, expectedKey.Public(), reparsedKey.Public())
-}
